Problem with resolver
