+++ src/app/catalog/catalog-routing.module.ts
@@ -1,6 +1,7 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CatalogComponent, EquipmentItemComponent } from '@app/catalog/containers';
+import { PetKindsResolver } from '@app/shared/resolver/pet-kinds.resolver';
 import { CategoriesComponent } from './containers/categories/categories.component';
 import { CategoriesResolver } from './resolver/categories.resolver';
 
@@ -10,12 +11,14 @@ const routes: Routes = [
     component: CatalogComponent,
     resolve: {
       activeCategories: CategoriesResolver,
+      petKinds: PetKindsResolver,
     },
   },
   {
     path: 'categories',
     resolve: {
       activeCategories: CategoriesResolver,
+      petKinds: PetKindsResolver,
     },
     children: [
       {
@@ -37,6 +40,6 @@ const routes: Routes = [
 @NgModule({
   imports: [RouterModule.forChild(routes)],
   exports: [RouterModule],
-  providers: [CategoriesResolver],
+  providers: [CategoriesResolver, PetKindsResolver],
 })
 export class CatalogRoutingModule {}
