+++ src/app/catalog/containers/catalog/catalog.component.html
@@ -1,7 +1,7 @@
-<lc-filter
+<!-- <lc-filter
   (outFilterValue)="receiveFilterValue($event)"
   (outSelectedFilters)="receiveSelectedFilters($event)"
-></lc-filter>
+></lc-filter> -->
 <lc-category-set></lc-category-set>
 <section class="catalog">
   <div class="catalog-search">
