+++ src/app/shared/components/filter-modal/filter-modal.component.html
@@ -1,4 +1,4 @@
-<!-- <div class="filter-modal" [formGroup]="filterGroup">
+<div class="filter-modal" [formGroup]="filterGroup">
   <div class="filter-modal__header">
     <h2 class="filter-modal__title">Фильтр</h2>
     <div (click)="closeModal()">
@@ -59,4 +59,4 @@
       </div>
     </div>
   </div>
-</div> -->
+</div>
