+++ src/app/shared/components/filter-modal/filter-modal.component.scss
