+++ src/app/shared/components/filter/filter.component.html
@@ -1,4 +1,4 @@
-<!-- <div class="filter">
+<div class="filter">
 <div *ngIf="selectedFilters > 0" matBadge="{{ selectedFilters }}"></div>
 <img src="assets/img/icons/icon-filter.svg" (click)="openModal()" class="svgfix" alt="filter_icon" />
-</div> -->
+</div>
