+++ src/app/shared/components/main-header/main-header.component.html
@@ -6,4 +6,5 @@
   <a routerLink="/auth">
     <img src="assets/img/icons/icon-avatar.svg" alt="avatar"/>
   </a>
+  <lc-filter *ngIf="pageTitle==='Каталог'"></lc-filter>
 </section>
