+++ src/app/shared/services/index.ts
@@ -1,0 +1,2 @@
+export { ApiService as CatalogApi } from './api/api.service';
+export { ControllerService as CatalogController } from './controller/controller.service';
